requirements.txt
pyOpenSSL
colorama
glob